+++ iCarousel No Nib Demo/iCarouselExampleViewController.h
@@ -1,0 +1,14 @@
+//
+//  iCarouselExampleViewController.h
+//  iCarouselExample
+//
+//  Created by Nick Lockwood on 03/04/2011.
+//  Copyright 2011 Charcoal Design. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+
+@interface iCarouselExampleViewController : UIViewController
+
+@end